Move list into MainWindowViewModel
