+++ Image-Sorter-UI/app/ViewModels/Interfaces/IFolderStructureDisplayViewModel.cs
@@ -1,5 +1,5 @@
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Reactive.Linq;
 using app.Model;
 using app.Views;
 
@@ -10,24 +10,17 @@ namespace app.ViewModels.Interfaces;
 /// </summary>
 public interface IFolderStructureDisplayViewModel
 {
-   /// <summary>
-   /// A list of <see cref="SelectFolders"/> used in choosing folder structure 
-   /// within the <see cref="FolderStructureDisplayView"/>
-   /// </summary>
-   public ObservableCollection<SelectFolders> FolderList { get; }
-   
+    /// <summary>
+    /// A List of folders
+    /// </summary>
+    public ObservableCollection<SelectFolders> FoldersList { get; }
+    
     /// <summary>
     /// Sets the <see cref="IMainWindowViewModel"/> context
     /// </summary>
     /// <param name="mainViewModel">An instance of <see cref="IMainWindowViewModel"/></param>
     public void SetMainViewModel(IMainWindowViewModel mainViewModel);
 
-    /// <summary>
-    /// Adds folders to <see cref="FolderList"/>
-    /// </summary>
-    /// <param name="folders">the folders being passed by the start processing button</param>
-    public void AddFolders(ObservableCollection<SelectFolders> folders);
-    
    /// <summary>
    /// Occurs when add images button is clicked returns to the
    /// <see cref="AddImageDisplayView"/>
